move: auth to util
